move knowledge of sparql-update from basicOperations to applyPatch in RDF component
